+++ tests/pipeline/test_gold_cartao.py
@@ -166,6 +166,7 @@ _SELECAO_FATO = (
     " +fact_despesa +fact_despesa_quarantine"
     " +fact_emenda"
     " +fact_cartao_cpgf +fact_cartao_cpgf_quarantine"
+    " +supplier_concentration +supplier_growth"
 )